[abieniusa, merge-from-trunk] Work in Progress: For the moment all files have the same length...
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/concordant/sources/swiftcloud/java/SwiftCloud-branches/1pc@1311 049c2a90-bdf1-4f83-8191-3719d539f8e0

+++ src/swift/application/filesystem/IFile.java
@@ -10,7 +10,7 @@ public interface IFile {
      * @param buf
      * @param offset
      */
-    void update(ByteBuffer buf, int offset);
+    void update(ByteBuffer buf, long offset);
 
     /**
      * Updates the content of the file with the data from the byte array.
@@ -25,7 +25,7 @@ public interface IFile {
      * @param buf
      * @param offset
      */
-    void read(ByteBuffer buf, int offset);
+    void read(ByteBuffer buf, long offset);
 
     /**
      * Returns content as byte array.
